CMSX 0.2.25 - Cleanup - Moved MSX version define to build tool - Added warning check for incompatible build flags - Added functions to set Input Manager timers - Fixed use of default (Main-ROM) font with text print - Added function to clear VRAM - Updated games build configuration (GoS & Tennis) - Updated samples build files (added a per sample configuration file) - Added a template project

+++ cmsx/src/compress.c
@@ -57,7 +57,7 @@ u16 UnpackRLEpToRAM(const u8* src, u8* dst RLEP_FIXSIZE_PARAM)
 		else if(type == 2) // Chunk of same 2 bytes
 		{
 			count <<= 1;
-			for(i8 i = 0; i < count; ++i)
+			for(u8 i = 0; i < count; ++i)
 				dst[i] = src[i & 0x1];
 			src += 2;
 		}
@@ -68,7 +68,7 @@ u16 UnpackRLEpToRAM(const u8* src, u8* dst RLEP_FIXSIZE_PARAM)
 		}
 		dst += count;
 	}
-	return (dst - start);
+	return ((u16)dst - start);
 }
 
 #endif // (USE_COMPRESS_RLEP)
